Section 3 - End of Section
WOO! :frog: :monkey: :horse: :racehorse: :camel: :sheep: :elephant: :panda_face: :snake: :bird:

That was a ride. I fucked up the database by adding in Genre as a string property on Movie before I knew that it was meant to be its own table with an Id and FK. I had to:
1. Delete Genre from Movie --> migrate
2. Add the Genre table --> migrate
3. Add GenreId field to Movie --> migrate
4. Add data to GenreId in Movie
5. Add Genre object to Movie --> migrate

Everything else is fine. :brain: